move email action to its own file, fix validation/parsing
